Improve recipe packaging script
This commit will improve the usability of the packaging script, especially for beginners.

It will output better error messages if there are any problems with the recipe. It will also add more tests to improve detecting those problems.
This way, more problems can be detected before even needing to create a PR.