Add support for empty unread badges
Empty unread badges show up when people manually "Mark as Unread" chat threads. Currently, this stops notifications from appearing as counts become NaN. With this PR they will stop being counted (since we don't have message counts in those cases and it's not really a notification I think this is probably the best behavior, but I'm ok with defaulting to 1).