More careful Gmail unread count detection
10a670fe33bd7fab841ddaf4674deb11477f2267 changed the unread count
detection to grab the number besides the Inbox link instead of the total
number of unread messages in the traditional inbox.
However, this detection may fail in there are no unread messages in the
inbox and report the number of unread messages from a different category
instead.

This commit combines the previous and the combined approach by first
finding the "Inbox" link as in the previous approach, but it extracts
the number of unread messages from the number beside it (instead of the
aria-label with the total number of unread messages).