Instance logo for Pleroma
Fetch instance logo from the instance configuration and the instance
colors from the pleroma-fe css to draw the instance logo.

If the instance logo is in full color, it will be shown on the instance
top bar background color.

If it is configured to display with a single color, alpha blending on a
<canvas> is used to render the logo in the top bar text color on top of
the top bar background color, as it appears on the webpage.