Fix badge for Zoho Mail
As `window.zmNCenter` which was used to get notifications count for the
badge's indirect notifications part does not exist anymore, badge for
Zoho Mail was not working.
The solution was to retrieve notifications count differently and to
exclude Zoho Mail specific `window` properties from the `if` condition
so the badge would be affected only partially if one of the used Zoho
Mail `window` properties gets removed. For example, if `zmTopBar` will
be removed from `window`, retrieval of the direct messages count would
still work.